Various style changes Remove `int` and `unsigned` in favor of `size_t` and `stdint.h` Remove explicit NULL compares Remove `fscanf`s in favor of custom parsing functions.

+++ src/include/registers.h
@@ -19,7 +19,7 @@ typedef struct CPUReg {
 typedef struct LocalVar {
     uint8_t type;
     // VArray of pointers to any occarance in which this local is referenced.
-    unsigned** references;
+    uint64_t** references;
     struct Statement* origin;
 } LocalVar;
